--- NoItemDrop/plugin.yml
@@ -1,4 +1,0 @@
-main: me.noodles.nodrop.Main
-name: NoDrop
-version: 0.1
-author: BGHDDevelopment(Noodles)
